+++ src/TestableHttpClient.NFluent/AssemblyInfo.cs
@@ -1,3 +1,6 @@
-﻿using System.Runtime.CompilerServices;
+﻿using System;
+using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("TestableHttpClient.NFluent.Tests")]
+
+[assembly: CLSCompliant(false)]