+++ src/TestableHttpClient/AssemblyInfo.cs
@@ -1,4 +1,7 @@
-﻿using System.Runtime.CompilerServices;
+﻿using System;
+using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("TestableHttpClient.NFluent")]
 [assembly: InternalsVisibleTo("TestableHttpClient.Tests")]
+
+[assembly: CLSCompliant(true)]